-Change structs to use int/long instead of uint/ulong -Update long references to point to either CORDB_ADDRESS or new CLRDATA_ADDRESS type where appropriate

+++ ManagedCorDebug/Managed/ResultTypes/GetDebugInfoResult.cs
@@ -1,14 +1,16 @@
+using System;
+
 namespace ManagedCorDebug
 {
     public struct GetDebugInfoResult
     {
-        public long PIDD { get; }
+        public IntPtr PIDD { get; }
 
         public int PcData { get; }
 
         public byte[] Data { get; }
 
-        public GetDebugInfoResult(long pIDD, int pcData, byte[] data)
+        public GetDebugInfoResult(IntPtr pIDD, int pcData, byte[] data)
         {
             PIDD = pIDD;
             PcData = pcData;
